# docs/openapi.yaml
components:
  schemas:
...
    DashboardDataQualityResponse:
      description: Dashboard data quality response.
      properties:
        assumed_label_count:
          title: Assumed Label Count
          type: integer
        exact_label_count:
          title: Exact Label Count
          type: integer
        low_confidence_count:
          title: Low Confidence Count
          type: integer
        low_confidence_items:
          items:
            $ref: '#/components/schemas/LowConfidenceItemResponse'
          title: Low Confidence Items
          type: array
        manual_count:
          title: Manual Count
          type: integer
        pattern_count:
          title: Pattern Count
          type: integer
        product_db_count:
          title: Product Db Count
          type: integer
        restaurant_db_count:
          title: Restaurant Db Count
          type: integer
        total_item_count:
          title: Total Item Count
          type: integer
        visual_estimate_count:
          title: Visual Estimate Count
          type: integer
      required:
      - exact_label_count
      - assumed_label_count
      - restaurant_db_count
      - product_db_count
      - pattern_count
      - visual_estimate_count
      - manual_count
      - low_confidence_count
      - total_item_count
      - low_confidence_items
      title: DashboardDataQualityResponse
      type: object
    DashboardDayResponse:
      description: Daily dashboard row.
      properties:
        carbs_g:
          title: Carbs G
          type: number
        date:
          format: date
          title: Date
          type: string
        fat_g:
          title: Fat G
          type: number
        fiber_g:
          title: Fiber G
          type: number
        kcal:
          title: Kcal
          type: number
        meal_count:
          title: Meal Count
          type: integer
        nutrients:
          items:
            $ref: '#/components/schemas/DashboardNutrientTotal'
          title: Nutrients
          type: array
        protein_g:
          title: Protein G
          type: number
      required:
      - date
      - kcal
      - carbs_g
      - protein_g
      - fat_g
      - fiber_g
      - meal_count
      title: DashboardDayResponse
      type: object
    DashboardHeatmapCell:
      description: Meal heatmap aggregate cell.
      properties:
        avg_carbs_g:
          title: Avg Carbs G
          type: number
        day_of_week:
          title: Day Of Week
          type: integer
        hour:
          title: Hour
          type: integer
        meal_count:
          title: Meal Count
          type: integer
      required:
      - day_of_week
      - hour
      - avg_carbs_g
      - meal_count
      title: DashboardHeatmapCell
      type: object
    DashboardHeatmapResponse:
      description: Dashboard meal heatmap response.
      properties:
        cells:
          items:
            $ref: '#/components/schemas/DashboardHeatmapCell'
          title: Cells
          type: array
      required:
      - cells
      title: DashboardHeatmapResponse
      type: object
    DashboardNutrientTotal:
      description: Daily or range nutrient total with known-value coverage.
      properties:
        amount:
          anyOf:
          - type: number
          - type: 'null'
          title: Amount
        coverage:
          title: Coverage
          type: number
        display_name:
          title: Display Name
          type: string
        known_item_count:
          title: Known Item Count
          type: integer
        nutrient_code:
          title: Nutrient Code
          type: string
        total_item_count:
          title: Total Item Count
          type: integer
        unit:
          title: Unit
          type: string
      required:
      - nutrient_code
      - display_name
      - unit
      - known_item_count
      - total_item_count
      - coverage
      title: DashboardNutrientTotal
      type: object
    DashboardRangeResponse:
      description: Dashboard range response.
      properties:
        days:
          items:
            $ref: '#/components/schemas/DashboardDayResponse'
          title: Days
          type: array
        summary:
          $ref: '#/components/schemas/DashboardRangeSummary'
      required:
      - days
      - summary
      title: DashboardRangeResponse
      type: object
    DashboardRangeSummary:
      description: Dashboard range aggregate summary.
      properties:
        avg_carbs_g:
          title: Avg Carbs G
          type: number
        avg_fat_g:
          title: Avg Fat G
          type: number
        avg_fiber_g:
          title: Avg Fiber G
          type: number
        avg_kcal:
          title: Avg Kcal
          type: number
        avg_protein_g:
          title: Avg Protein G
          type: number
        nutrients:
          items:
            $ref: '#/components/schemas/DashboardNutrientTotal'
          title: Nutrients
          type: array
        total_carbs_g:
          title: Total Carbs G
          type: number
        total_fat_g:
          title: Total Fat G
          type: number
        total_fiber_g:
          title: Total Fiber G
          type: number
        total_kcal:
          title: Total Kcal
          type: number
        total_meals:
          title: Total Meals
          type: integer
        total_protein_g:
          title: Total Protein G
          type: number
      required:
      - avg_kcal
      - avg_carbs_g
      - avg_protein_g
      - avg_fat_g
      - avg_fiber_g
      - total_meals
      - total_kcal
      - total_carbs_g
      - total_protein_g
      - total_fat_g
      - total_fiber_g
      title: DashboardRangeSummary
      type: object
    DashboardSourceBreakdownResponse:
      description: Dashboard source kind breakdown.
      properties:
        days:
          title: Days
          type: integer
        items:
          items:
            $ref: '#/components/schemas/DashboardSourceBreakdownRow'
          title: Items
          type: array
      required:
      - days
      - items
      title: DashboardSourceBreakdownResponse
      type: object
    DashboardSourceBreakdownRow:
      description: Source kind count row.
      properties:
        count:
          title: Count
          type: integer
        source_kind:
          $ref: '#/components/schemas/ItemSourceKind'
      required:
      - source_kind
      - count
      title: DashboardSourceBreakdownRow
      type: object
    DashboardTodayResponse:
      description: Dashboard summary for today.
      properties:
        carbs_g:
          title: Carbs G
          type: number
        date:
          format: date
          title: Date
          type: string
        fat_g:
          title: Fat G
          type: number
        fiber_g:
          title: Fiber G
          type: number
        hours_since_last_meal:
          anyOf:
          - type: number
          - type: 'null'
          title: Hours Since Last Meal
        kcal:
          title: Kcal
          type: number
        last_meal_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Meal At
        meal_count:
          title: Meal Count
          type: integer
        nutrients:
          items:
            $ref: '#/components/schemas/DashboardNutrientTotal'
          title: Nutrients
          type: array
        prev_week_avg_carbs:
          title: Prev Week Avg Carbs
          type: number
        prev_week_avg_kcal:
          title: Prev Week Avg Kcal
          type: number
        protein_g:
          title: Protein G
          type: number
        week_avg_carbs:
          title: Week Avg Carbs
          type: number
        week_avg_kcal:
          title: Week Avg Kcal
          type: number
      required:
      - date
      - kcal
      - carbs_g
      - protein_g
      - fat_g
      - fiber_g
      - meal_count
      - week_avg_carbs
      - week_avg_kcal
      - prev_week_avg_carbs
      - prev_week_avg_kcal
      title: DashboardTodayResponse
      type: object
    DashboardTodayWithGlucoseResponse:
      description: Dashboard summary for a role with glucose context enabled.
      properties:
        carbs_g:
          title: Carbs G
          type: number
        current_glucose:
          anyOf:
          - type: number
          - type: 'null'
          title: Current Glucose
        current_glucose_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Current Glucose At
        date:
          format: date
          title: Date
          type: string
        fat_g:
          title: Fat G
          type: number
        fiber_g:
          title: Fiber G
          type: number
        hours_since_last_meal:
          anyOf:
          - type: number
          - type: 'null'
          title: Hours Since Last Meal
        kcal:
          title: Kcal
          type: number
        last_meal_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Meal At
        meal_count:
          title: Meal Count
          type: integer
        nutrients:
          items:
            $ref: '#/components/schemas/DashboardNutrientTotal'
          title: Nutrients
          type: array
        prev_week_avg_carbs:
          title: Prev Week Avg Carbs
          type: number
        prev_week_avg_kcal:
          title: Prev Week Avg Kcal
          type: number
        protein_g:
          title: Protein G
          type: number
        week_avg_carbs:
          title: Week Avg Carbs
          type: number
        week_avg_kcal:
          title: Week Avg Kcal
          type: number
      required:
      - date
      - kcal
      - carbs_g
      - protein_g
      - fat_g
      - fiber_g
      - meal_count
      - week_avg_carbs
      - week_avg_kcal
      - prev_week_avg_carbs
      - prev_week_avg_kcal
      - current_glucose
      - current_glucose_at
      title: DashboardTodayWithGlucoseResponse
      type: object
    DashboardTopPatternResponse:
      description: Top used pattern response row.
      properties:
        count:
          title: Count
          type: integer
        display_name:
          title: Display Name
          type: string
        pattern_id:
          format: uuid
          title: Pattern Id
          type: string
        token:
          title: Token
          type: string
      required:
      - pattern_id
      - token
      - display_name
      - count
      title: DashboardTopPatternResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    LowConfidenceItemResponse:
      description: Low-confidence item row.
      properties:
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        item_id:
          format: uuid
          title: Item Id
          type: string
        meal_id:
          format: uuid
          title: Meal Id
          type: string
        name:
          title: Name
          type: string
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
      required:
      - meal_id
      - item_id
      - name
      title: LowConfidenceItemResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  summary: Informational food diary API.
  title: Glucotracker API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /dashboard/data_quality:
    get:
      description: Return item source and confidence quality metrics.
      operationId: getDashboardDataQuality
      parameters:
      - in: query
        name: days
        required: false
        schema:
          default: 7
          maximum: 365
          minimum: 1
          title: Days
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardDataQualityResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Dashboard Data Quality
      tags:
      - dashboard
      x-glucotracker-scoped: true
  /dashboard/heatmap:
    get:
      description: Aggregate accepted meals by weekday and hour.
      operationId: getDashboardHeatmap
      parameters:
      - in: query
        name: weeks
        required: false
        schema:
          default: 4
          maximum: 52
          minimum: 1
          title: Weeks
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardHeatmapResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Dashboard Heatmap
      tags:
      - dashboard
      x-glucotracker-scoped: true
  /dashboard/range:
    get:
      description: Return daily dashboard rows and summary for an inclusive range.
      operationId: getDashboardRange
      parameters:
      - in: query
        name: from
        required: true
        schema:
          format: date
          title: From
          type: string
      - in: query
        name: to
        required: true
        schema:
          format: date
          title: To
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardRangeResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Dashboard Range
      tags:
      - dashboard
      x-glucotracker-scoped: true
  /dashboard/source_breakdown:
    get:
      description: Count accepted meal items by source kind.
      operationId: getDashboardSourceBreakdown
      parameters:
      - in: query
        name: days
        required: false
        schema:
          default: 7
          maximum: 365
          minimum: 1
          title: Days
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSourceBreakdownResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Dashboard Source Breakdown
      tags:
      - dashboard
      x-glucotracker-scoped: true
  /dashboard/today:
    get:
      description: Return today's dashboard summary.
      operationId: getDashboardToday
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/DashboardTodayWithGlucoseResponse'
                - $ref: '#/components/schemas/DashboardTodayResponse'
                title: Response Getdashboardtoday
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Dashboard Today
      tags:
      - dashboard
      x-glucotracker-food-omits:
      - glucose
      - nightscout
      x-glucotracker-role-variant: true
      x-glucotracker-scoped: true
  /dashboard/top_patterns:
    get:
      description: Return most frequently used patterns in accepted meals.
      operationId: getDashboardTopPatterns
      parameters:
      - in: query
        name: days
        required: false
        schema:
          default: 7
          maximum: 365
          minimum: 1
          title: Days
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DashboardTopPatternResponse'
                title: Response Getdashboardtoppatterns
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - BearerAuth: []
      summary: Dashboard Top Patterns
      tags:
      - dashboard
      x-glucotracker-scoped: true
...